+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Now, instead, the "Executor" classes implement the `getBinary()` method, which r
   `SqliteExecutor` now explicitly return `void`;
 * the `AbstractBackupUtility::__get()` magic method, which allowed reading the properties of the "Executor" classes, had
   been deprecated and has now been removed;
+* migration to Mockery. We need to use the `TestCase` class of the plugin;
 * the `DatabaseBackup.connection` configuration name had been deprecated and now have been removed;
 * configuration names `DatabaseBackup.mysql`, `DatabaseBackup.postgres` and `DatabaseBackup.sqlite` had been deprecated, 
   and now they have been removed;